backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
supabase==1.0.0
python-jose==3.3.0
PyJWT==2.8.0
scikit-learn==1.4.2
pandas==2.2.3
pytest==7.4.0
requests==2.31.0
python-dotenv==1.0.0
